+++ src/categories/js/eslint/config/react-ts.config.ts
