+++ src/categories/js/jest/jest.config.ts
@@ -1,7 +1,7 @@
 import { jsCategoryState } from '#src/states/categories.ts';
 import { defaultConfig } from '#src/categories/js/jest/config/default.config.ts';
-import { nodeTsConfig } from '#src/categories/js/jest/config/node:ts.config.ts';
-import { reactTsConfig } from '#src/categories/js/jest/config/react:ts.config.ts';
+import { nodeTsConfig } from '#src/categories/js/jest/config/node-ts.config.ts';
+import { reactTsConfig } from '#src/categories/js/jest/config/react-ts.config.ts';
 
 export const [getConfig] = jsCategoryState.useConfigState({
   default: defaultConfig,
