+++ src/categories/js/lint-staged/config/node-ts.config.ts
