+++ src/categories/js/lint-staged/config/react-ts.config.ts
