+++ src/categories/js/prettier/prettier.const.ts
@@ -1,4 +1,4 @@
-export const PACKAGE_NAME = 'prettier';
+export const PACKAGE_NAME = "prettier";
 export const CLI_NAME = PACKAGE_NAME;
-export const CONFIG_FILE_NAME = '.prettierrc';
-export const CONFIG_IGNORE_FILE_NAME = '.prettierignore';
+export const CONFIG_FILE_NAME = ".prettierrc";
+export const CONFIG_IGNORE_FILE_NAME = ".prettierignore";
