+++ src/categories/js/stylelint/config/react-ts.config.ts
